% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-seurat.R
\name{add_dummy_feature}
\alias{add_dummy_feature}
\title{Add a dummy feature to a SingleCellExperiment object with all zeros}
\usage{
add_dummy_feature(sce, feature_name = "null-feature")
}
\arguments{
\item{sce}{A SingleCellExperiment object. Should contain only one row/feature.
If larger, it will be returned unmodified.}

\item{feature_name}{The name to give to the new dummy feature}
}
\value{
A SingleCellExperiment object with a dummy feature added (all zeros)
}
\description{
Add a dummy feature to a SingleCellExperiment object with all zeros
}
