% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-clusters.R
\name{calculate_clusters}
\alias{calculate_clusters}
\title{Calculate graph-based clusters from a provided matrix}
\usage{
calculate_clusters(
  mat,
  algorithm = c("louvain", "walktrap", "leiden"),
  weighting = c("jaccard", "rank"),
  nn = 10,
  resolution = 1,
  objective_function = "CPM",
  cluster_args = list(),
  seed = NULL
)
}
\arguments{
\item{mat}{Matrix, usually of PCs, where each row is a cell. Matrix must have rownames of cell ids (e.g., barcodes)}

\item{algorithm}{Clustering algorithm to use. Must be one of "louvain" (default), "walktrap", or "leiden"}

\item{weighting}{Weighting scheme to use. Must be one of "jaccard" (default) or "rank"}

\item{nn}{Number of nearest neighbors. Default is 10.}

\item{resolution}{Resolution parameter used by louvain and leiden clustering only. Default is 1.}

\item{objective_function}{Leiden-specific parameter for whether to use the Constant Potts Model ("CPM"; default) or "modularity"}

\item{cluster_args}{List of additional arguments to pass to the clustering function.}

\item{seed}{Random seed to set for clustering. Default is 2024.}
}
\value{
A data frame of cluster results with columns `cell_id` and `cluster`. Additional columns represent algorithm parameters
  and include at least: `algorithm`, `weighting`, and `nn`. Louvain and leiden clustering will also include `resolution`, and
  leiden clustering will further include `objective_function`. Any other values passed to `cluster_args` will also be included.
}
\description{
Calculate graph-based clusters from a provided matrix
}
\examples{
\dontrun{
# TODO WILL ADD EXAMPLES ONCE ARGS ARE FINALIZED
}
}
