% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep-parameters.R
\name{calculate_clusters_sweep}
\alias{calculate_clusters_sweep}
\title{Calculate clusters across a set of parameters}
\usage{
calculate_clusters_sweep(
  x,
  algorithm = c("louvain", "walktrap", "leiden"),
  weighting = "jaccard",
  nn = 10,
  resolution = 1,
  objective_function = "CPM",
  cluster_args = list(),
  threads = 1,
  seed = NULL,
  pc_name = NULL
)
}
\arguments{
\item{x}{An object containing PCs that clustering can be performed in. This can be either a SingleCellExperiment
object, a Seurat object, or a matrix where columns are PCs and rows are cells. If a matrix is provided, it must
have row names of cell ids (e.g., barcodes).#' @param algorithm}

\item{algorithm}{Clustering algorithm to use. Must be one of "louvain" (default), "walktrap", or "leiden".}

\item{weighting}{Weighting scheme(s) to consider when sweeping parameters.
Provide a vector of unique values to vary this parameter. Options include  "jaccard" (default), "rank", or "number"}

\item{nn}{Number of nearest neighbors to consider when sweeping parameters.
Provide a vector of unique values to vary this parameter. Default is 10.}

\item{resolution}{Resolution parameter used by louvain and leiden clustering only.
Provide a vector of unique values to vary this parameter. Default is 1.}

\item{objective_function}{Leiden-specific parameter for whether to use the Constant Potts Model ("CPM"; default) or "modularity".
Provide a vector of unique values to vary this parameter.}

\item{cluster_args}{List of additional arguments to pass to the chosen clustering function.
Parameters values in this list cannot be varied.
Only single values for each argument are supported (no vectors or lists).
See igraph documentation for details on each clustering function: https://igraph.org/r/html/latest}

\item{threads}{Number of threads to use. Default is 1.}

\item{seed}{Random seed to set for clustering.}

\item{pc_name}{Name of principal components slot in provided object. This argument is only used if a SingleCellExperiment
or Seurat object is provided. If not provided, the SingleCellExperiment object name will default to "PCA" and the
Seurat object name will default to "pca".}
}
\value{
A data frame with results from performing clustering with all parameter combinations.
  Columns include `cluster_set` (identifier column for results from a single clustering run),
  `cell_id`, and `cluster`. Additional columns represent algorithm parameters and include at least:
  `algorithm`, `weighting`, and `nn`. Louvain and leiden clustering will also include `resolution`,
  and leiden clustering will further include `objective_function`.
}
\description{
This function can be used to perform reproducible clustering while varying a set of parameters.
A single clustering algorithm is required, but multiple values can be provided for any of:
 - The weighting scheme (`weighting`)
 - Number of nearest neighrbors (`nn`)
 - The resolution parameter (`resolution`)
 - The objective function parameter (`objective_function`)
Note that defaults for some arguments may differ from the bluster::NNGraphParam() defaults.
Specifically, the clustering algorithm defaults to "louvain" and the weighting scheme to "jaccard"
to align with common practice in scRNA-seq analysis.
}
\examples{
\dontrun{
# performing louvain clustering with jaccard weighting,
# varying the nearest neighobor parameter.
cluster_df <- calculate_clusters_sweep(sce_object, nn = c(10, 15, 20, 25))

# performing louvain clustering, with jaccard and rank weighting, and
# varying the nearest neighbor and resolution parameters.
cluster_df <- calculate_clusters_sweep(
  sce_object,
  weighting = c("jaccard", "rank"),
  nn = c(10, 15, 20, 25),
  resolution = c(0.5, 1)
)
}
}
