% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-clusters.R
\name{calculate_silhouette}
\alias{calculate_silhouette}
\title{Calculate the silhouette width of clusters}
\usage{
calculate_silhouette(
  x,
  cluster_df,
  cluster_col = "cluster",
  cell_id_col = "cell_id",
  pc_name = NULL
)
}
\arguments{
\item{x}{Either a matrix of principal components (PCs), or a SingleCellExperiment
or Seurat object containing PCs. If a matrix is provided, rows should be cells
and columns should be PCs, and row names should be cell ids (e.g., barcodes).}

\item{cluster_df}{A data frame that contains at least two columns: one representing
unique cell ids, and one containing cluster assignments. By default, these columns
should be named `cell_id` and `cluster` respectively, though this can be customized.
The cell id column's values should match either the PC matrix row names, or the
SingleCellExperiment/Seurat object cell ids. Typically this data frame will be
output from the `rOpenScPCA::calculate_clusters()` function.}

\item{cluster_col}{The name of the column in `cluster_df` which contains cluster
assignments. Default value is "cluster".}

\item{cell_id_col}{The name of the column in `cluster_df` which contains unique cell
ids "cell_id" is assumed by default.}

\item{pc_name}{Optionally, the name of the PC matrix in the object. Not used if a
matrix is provided. If the name is not provided, the name "PCA" is assumed for
SingleCellExperiment objects, and "pca" for Seurat objects.}
}
\value{
Expanded `cluster_df` data frame with additional columns `silhouette_width`,
  the cell's silhouette width, and `silhouette_other`, the closest cluster other
  than the one to which the given cell was assigned. For more information,
  see documentation for `bluster::approxSilhouette()`.
}
\description{
This function uses the `bluster::approxSilhouette()` function to calculate the
silhouette width for a clustering result. These results can be used downstream to
calculate the average silhouette width, a popular metric for cluster evaluation.
}
\examples{
\dontrun{
# calculate silhouette width for clusters stored in a data frame
cluster_df <- calculate_silhouette(sce_object, cluster_df)
}
}
