% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-clusters.R
\name{extract_pc_matrix}
\alias{extract_pc_matrix}
\title{Extract a principal components (PC) matrix from either a SingleCellExperiment
or a Seurat object.}
\usage{
extract_pc_matrix(sc_object, pc_name = NULL)
}
\arguments{
\item{sc_object}{Either a SingleCellExperiment or Seurat object}

\item{pc_name}{Optionally, the name of the PC matrix in the object. If this is
not provided, the name "PCA" is assumed for SingleCellExperiment objects, and
"pca" for Seurat objects.}
}
\value{
PC matrix with row names
}
\description{
This function first determines if the provided object is a SingleCellExperiment or
Seurat object, and then extract the PC matrix. If no name for the PC matrix is provided,
this function will assume the name of "PCA" for SingleCellExperiment objects, and
"pca" for Seurat objects.
}
