% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-clusters.R
\name{prepare_pc_matrix}
\alias{prepare_pc_matrix}
\title{Helper function to check and/or extract a matrix of PCs from a given object}
\usage{
prepare_pc_matrix(x, pc_name = NULL)
}
\arguments{
\item{x}{Either a matrix of principal components (PCs), or a SingleCellExperiment
or Seurat object containing PCs. If a matrix is provided, rows should be cells
and columns should be PCs, and row names should be cell ids (e.g., barcodes).}

\item{pc_name}{Optionally, the name of the PC matrix in the object. Not used for
matrices. If this is not provided, the name "PCA" is assumed for
SingleCellExperiment objects, and "pca" for Seurat objects.}
}
\value{
A matrix of PCs with row names representing cell ids
}
\description{
Helper function to check and/or extract a matrix of PCs from a given object
}
