% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-seurat.R
\name{sce_to_seurat}
\alias{sce_to_seurat}
\title{Convert an SCE object to Seurat}
\usage{
sce_to_seurat(
  sce,
  use_symbols = TRUE,
  reference = c("sce", "scpca", "10x2020", "10x2024"),
  dedup_method = c("unique", "sum"),
  seurat_assay_version = c("v3", "v5")
)
}
\arguments{
\item{sce}{The input SCE object}

\item{use_symbols}{Logical indicating whether the Seurat object uses gene
symbols for indexing. Default is TRUE.}

\item{reference}{If `use_symbols` is TRUE, the reference to use for gene
symbols. One of `sce`, `scpca`, `10x2020`, `10x2024`, where `sce` uses the
symbols stored in the `gene_symbol` column of the SCE object's row data,
and the others use standardized translation tables. See `ensembl_to_symbol()`
for more information. Default is `sce`.}

\item{dedup_method}{Method to handle duplicated gene symbols. If `unique`,
the gene symbols will be made unique following standard Seurat procedures.
If `sum`, the expression values for each gene symbols will be summed.}

\item{seurat_assay_version}{Whether to create Seurat `v3` or `v5` assay
objects. Default is `v3`.}
}
\value{
a Seurat object
}
\description{
Converts an ScPCA SingleCellExperiment (SCE) object to Seurat format. This is
primarily a wrapper around Seurat::as.Seurat() with some additional steps to
include ScPCA metadata and options for converting the feature index from
Ensembl gene ids to gene symbols.
}
\details{
If present, reduced dimensions from SCE objects will be included, renamed to
match Seurat default naming.
}
\examples{
\dontrun{
# convert an ScPCA SCE to Seurat renaming with gene symbols
seurat_obj <- sce_to_seurat(sce)

# convert SCE to Seurat keeping the current names (usually Ensembl)
seurat_obj <- sce_to_seurat(sce, use_symbols = FALSE)

# convert SCE to Seurat, merging duplicated gene names
seurat_obj <- sce_to_seurat(sce, dedup_method = "sum")

# convert SCE to Seurat with v5 assay objects
seurat_obj <- sce_to_seurat(sce, seurat_assay_version = "v5")
}

}
