% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-gene-ids.R
\name{sce_to_symbols}
\alias{sce_to_symbols}
\title{Set the row names of an ScPCA SingleCellExperiment object to gene symbols}
\usage{
sce_to_symbols(sce, convert_hvg = TRUE, convert_pca = TRUE)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing gene ids and gene symbols.}

\item{convert_hvg}{Logical indicating whether to convert highly variable genes to gene symbols.}

\item{convert_pca}{Logical indicating whether to convert PCA rotation matrix to gene symbols.}
}
\value{
A SingleCellExperiment object with row names set as gene symbols.
}
\description{
The SingleCellExperiment objects produced as part of ScPCA are indexed by
Ensembl gene ids, as those are more stable than gene symbols. However,
for many applications gene symbols are useful. This function converts the
row names (indexes) of a SingleCellExperiment object to gene symbols based.
}
\details{
Internal data structures such as the list of highly variable genes and the
rotation matrix for the PCA are also updated to use gene symbols, if present
(and not disabled by the `convert_hvg` and `convert_pca` arguments).

Note that using this function will result in non-unique row ids as no
de-duplication is currently performed.
}
\examples{
\dontrun{
# convert a SingleCellExperiment object to use gene symbols
symbol_sce <- sce_to_symbols(sce)
}
}
