% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum-duplicate-genes.R
\name{sum_duplicate_genes}
\alias{sum_duplicate_genes}
\title{Sum counts for genes with duplicate names in a SingleCellExperiment object.}
\usage{
sum_duplicate_genes(sce, normalize = TRUE, recalculate_reduced_dims = FALSE)
}
\arguments{
\item{sce}{a SingleCellExperiment object with duplicated row names}

\item{normalize}{a logical indicating whether to normalize the expression
values. Default is TRUE}

\item{recalculate_reduced_dims}{a logical indicating whether to recalculate
PCA and UMAP. If FALSE, the input reduced dimensions are copied over. If
TRUE, the highly variable genes are also recalculated with the new values
stored in metadata. Default is FALSE}
}
\value{
a SingleCellExperiment object
}
\description{
Genes with the same name are merged by summing their raw expression counts.
If requested, the lognormalized expression values are recalculated, otherwise
this is left blank.
}
