% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep-clusters.R
\name{sweep_clusters}
\alias{sweep_clusters}
\title{Calculate clusters across a set of parameters}
\usage{
sweep_clusters(
  x,
  algorithm = "louvain",
  weighting = "jaccard",
  nn = 10,
  resolution = 1,
  objective_function = "CPM",
  threads = 1,
  seed = NULL,
  pc_name = NULL
)
}
\arguments{
\item{x}{An object containing PCs that clustering can be performed in. This can be either
a SingleCellExperiment object, a Seurat object, or a matrix where columns are PCs and
rows are cells. If a matrix is provided, it must have row names of cell ids (e.g., barcodes).}

\item{algorithm}{Clustering algorithm to use. Must be one of "louvain" (default), "walktrap",
or "leiden".}

\item{weighting}{Weighting scheme(s) to consider when sweeping parameters.
Provide a vector of unique values to vary this parameter. Options include "jaccard" (default),
  "rank", or "number"}

\item{nn}{Number of nearest neighbors to consider when sweeping parameters.
Provide a vector of unique values to vary this parameter. Default is 10.}

\item{resolution}{Resolution parameter used by louvain and leiden clustering only.
Provide a vector of unique values to vary this parameter. Default is 1.}

\item{objective_function}{Leiden-specific parameter for whether to use the
Constant Potts Model ("CPM"; default) or "modularity". Provide a vector of unique values
to vary this parameter.}

\item{threads}{Number of threads to use. Default is 1.}

\item{seed}{Random seed to set for clustering.}

\item{pc_name}{Name of principal components slot in provided object. This argument is only used
if a SingleCellExperiment or Seurat object is provided. If not provided, the SingleCellExperiment
object name will default to "PCA" and the Seurat object name will default to "pca".}
}
\value{
A list of data frames from performing clustering across all parameter combinations.
  Columns include `cluster_set` (identifier column for results from a single clustering run),
  `cell_id`, and `cluster`. Additional columns represent algorithm parameters and include at least:
  `algorithm`, `weighting`, and `nn`. Louvain and leiden clustering will also include `resolution`,
  and leiden clustering will further include `objective_function`.
}
\description{
This function can be used to perform reproducible clustering while varying a set of parameters.
Multiple values can be provided for any of:
 - The algorithm (`algorithm`)
 - The weighting scheme (`weighting`)
 - Number of nearest neighbors (`nn`)
 - The resolution parameter (`resolution`)
 - The objective function parameter (`objective_function`)
}
\details{
For each algorithm specified, all parameters possible to use with that
algorithm will be systematically varied. This function does not accept additional
parameters besides those listed above.
Note that defaults for some arguments may differ from the bluster::NNGraphParam() defaults.
Specifically, the clustering algorithm defaults to "louvain" and the weighting scheme to "jaccard"
to align with common practice in scRNA-seq analysis.
}
\examples{
\dontrun{
# perform louvain clustering with jaccard weighting (defaults),
# varying the nearest neighobor parameter, and set a seed for reproducibility
cluster_df <- sweep_clusters(
  sce_object,
  nn = c(10, 15, 20, 25),
  seed = 11
 )

# perform louvain clustering, with jaccard and rank weighting, and
# varying the nearest neighbor and resolution parameters.
cluster_df <- sweep_clusters(
  sce_object,
  algorithm = "louvain",
  weighting = c("jaccard", "rank"),
  nn = c(10, 15, 20, 25),
  resolution = c(0.5, 1),
  seed = 11
)

# perform walktrap and louvain clustering with jaccard weighting, and
# varying the nearest neighbors for both algorithms, and resolution for louvain.
cluster_df <- sweep_clusters(
  sce_object,
  algorithm = c("walktrap", "louvain"),
  weighting = "jaccard",
  nn = c(10, 15, 20, 25),
  resolution = c(0.5, 1),
  seed = 11
)
}
}
