% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-clusters.R
\name{calculate_clusters}
\alias{calculate_clusters}
\title{Calculate graph-based clusters from a provided matrix}
\usage{
calculate_clusters(
  mat,
  algorithm = c("louvain", "walktrap", "leiden"),
  weighting = c("jaccard", "rank", "number"),
  nn = 10,
  resolution = 1,
  objective_function = c("CPM", "modularity"),
  cluster_args = list(),
  seed = NULL
)
}
\arguments{
\item{mat}{Matrix, usually of PCs, where each row is a cell. Matrix must have rownames of cell ids (e.g., barcodes)}

\item{algorithm}{Clustering algorithm to use. Must be one of "louvain" (default), "walktrap", or "leiden".}

\item{weighting}{Weighting scheme to use. Must be one of "jaccard" (default), "rank", or "number"}

\item{nn}{Number of nearest neighbors. Default is 10.}

\item{resolution}{Resolution parameter used by louvain and leiden clustering only. Default is 1.}

\item{objective_function}{Leiden-specific parameter for whether to use the Constant Potts Model ("CPM"; default) or "modularity"}

\item{cluster_args}{List of additional arguments to pass to the chosen clustering function. Only single-length values will be used.
See igraph documentation for details on each clustering function: https://igraph.org/r/html/latest}

\item{seed}{Random seed to set for clustering. Default is 2024.}
}
\value{
A data frame of cluster results with columns `cell_id` and `cluster`. Additional columns represent algorithm parameters
  and include at least: `algorithm`, `weighting`, and `nn`. Louvain and leiden clustering will also include `resolution`, and
  leiden clustering will further include `objective_function`.
}
\description{
This function is provided to simplify application of bluster package clustering functions on OpenScPCA data.
In particular, this function runs bluster::clusterRows() with the bluster::NNGraphParam() function.
Note that defaults for some arguments may differ from the bluster::NNGraphParam() defaults.
Specifically, the clustering algorithm defaults to "louvain" and the distance measure to "jaccard"
to align with common practice in scRNA-seq anaylsis.
}
\examples{
\dontrun{
# cluster using default parameters
cluster_df <- calculate_clusters(pca_matrix)

# cluster using the leiden algorithm with a resolution of 0.1
cluster_df <- calculate_clusters(pca_matrix, algorithm = "leiden", resolution = 0.1)

# cluster using the leiden algorithm with a non-default of 3 iterations
cluster_df <- calculate_clusters(
  pca_matrix,
  algorithm = "leiden",
  cluster_args = list(n_iterations = 3)
)
}
}
