% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-clusters.R
\name{calculate_clusters}
\alias{calculate_clusters}
\title{Calculate graph-based clusters from a provided matrix}
\usage{
calculate_clusters(
  x,
  algorithm = c("louvain", "walktrap", "leiden"),
  weighting = c("jaccard", "rank", "number"),
  nn = 10,
  resolution = 1,
  objective_function = c("CPM", "modularity"),
  cluster_args = list(),
  threads = 1,
  seed = NULL,
  pc_name = NULL
)
}
\arguments{
\item{x}{An object containing PCs that clustering can be performed in. This can be either a SingleCellExperiment
object, a Seurat object, or a matrix where columns are PCs and rows are cells. If a matrix is provided, it must
have row names of cell ids (e.g., barcodes).}

\item{algorithm}{Clustering algorithm to use. Must be one of "louvain" (default), "walktrap", or "leiden".}

\item{weighting}{Weighting scheme to use. Must be one of "jaccard" (default), "rank", or "number"}

\item{nn}{Number of nearest neighbors. Default is 10.}

\item{resolution}{Resolution parameter used by louvain and leiden clustering only. Default is 1.}

\item{objective_function}{Leiden-specific parameter for whether to use the Constant Potts Model ("CPM"; default) or "modularity"}

\item{cluster_args}{List of additional arguments to pass to the chosen clustering function.
Only single values for each argument are supported (no vectors or lists).
See igraph documentation for details on each clustering function: https://igraph.org/r/html/latest}

\item{threads}{Number of threads to use. Default is 1.}

\item{seed}{Random seed to set for clustering.}

\item{pc_name}{Name of principal components slot in provided object. This argument is only used if a SingleCellExperiment
or Seurat object is provided. If not provided, the SingleCellExperiment object name will default to "PCA" and the
Seurat object name will default to "pca".}
}
\value{
A data frame of cluster results with columns `cell_id` and `cluster`. Additional columns represent algorithm parameters
  and include at least: `algorithm`, `weighting`, and `nn`. Louvain and leiden clustering will also include `resolution`, and
  leiden clustering will further include `objective_function`.
}
\description{
This function is provided to simplify application of bluster package clustering functions on OpenScPCA data.
In particular, this function runs bluster::clusterRows() with the bluster::NNGraphParam() function on a
principal components matrix, provided either directly or via single-cell object.
Note that defaults for some arguments may differ from the bluster::NNGraphParam() defaults.
Specifically, the clustering algorithm defaults to "louvain" and the weighting scheme to "jaccard"
to align with common practice in scRNA-seq analysis.
}
\examples{
\dontrun{
# cluster PCs from a SingleCellExperiment object using default parameters and
# a random seed for reproducibility
cluster_df <- calculate_clusters(sce_object, seed = 11)

# cluster PCs from a SingleCellExperiment object using default parameters and 4 threads
cluster_df <- calculate_clusters(sce_object, threads = 4, seed = 11)

# cluster PCs from a Seurat object using default parameters
cluster_df <- calculate_clusters(seurat_object, seed = 11)

# cluster directly from a matrix using default parameters
cluster_df <- calculate_clusters(pca_matrix, seed = 11)

# cluster directly from a matrix using the leiden algorithm with a resolution of 0.1
cluster_df <- calculate_clusters(
  pca_matrix,
  algorithm = "leiden",
  resolution = 0.1,
  seed = 11
 )

# cluster directly from a matrix using the leiden algorithm with 3 iterations
cluster_df <- calculate_clusters(
  pca_matrix,
  algorithm = "leiden",
  cluster_args = list(n_iterations = 3),
  seed = 11
)
}
}
