% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-clusters.R
\name{calculate_purity}
\alias{calculate_purity}
\title{Calculate the neighborhood purity of clusters}
\usage{
calculate_purity(x, cluster_df, pc_name = NULL, ...)
}
\arguments{
\item{x}{Either a matrix of principal components (PCs), or a SingleCellExperiment
or Seurat object containing PCs. If a matrix is provided, rows should be cells
and columns should be PCs, and row names should be cell ids (e.g., barcodes).}

\item{cluster_df}{A data frame that contains at least the columns `cell_id` and
`cluster`. The `cell_id` values should match either the PC matrix row names,
or the SingleCellExperiment/Seurat object cell ids. Typically this will be output from
the `rOpenScPCA::calculate_clusters()` function.}

\item{pc_name}{Optionally, the name of the PC matrix in the object. Not used if a
matrix is provided. If the name is not provided, the name "PCA" is assumed for
SingleCellExperiment objects, and "pca" for Seurat objects.}

\item{...}{Additional arguments to pass to `bluster::neighborPurity()`}
}
\value{
Expanded `cluster_df` data frame with the additional column `purity`
}
\description{
This function uses the `bluster::neighborPurity()` function to calculate the
neighborhood purity values for a clustering result.
}
\examples{
\dontrun{
# calculate neighborhood purity for clusters stored in a data frame
cluster_df <- calculate_purity(sce_object, cluster_df)
}
}
