% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-clusters.R
\name{calculate_silhouette}
\alias{calculate_silhouette}
\title{Calculate the silhouette width of clusters}
\usage{
calculate_silhouette(x, cluster_df, pc_name = NULL)
}
\arguments{
\item{x}{Either a matrix of principal components (PCs), or a SingleCellExperiment
or Seurat object containing PCs. If a matrix is provided, rows should be cells
and columns should be PCs, and row names should be cell ids (e.g., barcodes).}

\item{cluster_df}{A data frame that contains at least the columns `cell_id` and
`cluster`. The `cell_id` values should match either the PC matrix row names,
or the SingleCellExperiment/Seurat object cell ids. Typically this will be output from
the `rOpenScPCA::calculate_clusters()` function.}

\item{pc_name}{Optionally, the name of the PC matrix in the object. Not used if a
matrix is provided. If the name is not provided, the name "PCA" is assumed for
SingleCellExperiment objects, and "pca" for Seurat objects.}
}
\value{
Expanded `cluster_df` data frame with the additional column `silhouette_width`
}
\description{
This function uses the `bluster::approxSilhouette()` function to calculate the
silhouette width for a clustering result. These results can be used downstream to
calculate the average silhouette width, a popular metric for cluster evaluation.
}
\examples{
\dontrun{
# calculate silhouette width for clusters stored in a data frame
cluster_df <- calculate_silhouette(sce_object, cluster_df)
}
}
