% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-gene-ids.R
\name{ensembl_to_symbol}
\alias{ensembl_to_symbol}
\title{Convert Ensembl gene ids to gene symbols based on an ScPCA SingleCellExperiment object}
\usage{
ensembl_to_symbol(ensembl_ids, sce, leave_na = FALSE)
}
\arguments{
\item{ensembl_ids}{A character vector of Ensembl gene ids to translate to
gene symbols.}

\item{sce}{A SingleCellExperiment object containing gene ids and gene symbols
to use for translation.}

\item{leave_na}{logical indicating whether to leave NA values in the output.
Default is `FALSE`}
}
\value{
A vector of gene symbols corresponding to the input Ensembl ids.
}
\description{
The SingleCellExperiment objects produced as part of ScPCA are indexed by
Ensembl gene ids, as those are more stable than gene symbols. However,
for many applications gene symbols are useful. This function provides a
simple and consistent conversion of Ensembl gene ids to gene symbols based on
the `gene_symbol` column that is present in the row data of ScPCA
SingleCellExperiment objects.
}
\details{
For this function, the SingleCellExperiment object must contain a `gene_ids`
column containing Ensembl gene ids and a `gene_symbol` column containing gene
symbols. If any gene ids are not found or if the gene symbol is not defined,
the input gene id is returned, unless the `leave_na` is set to `TRUE`.
}
\examples{
\dontrun{
# convert a set of Ensembl ids to gene symbols
# using a SingleCellExperiment reference
ensembl_ids <- c("ENSG00000141510", "ENSG00000134323")
gene_symbols <- ensembl_to_symbol(ensembl_ids, sce)
gene_symbols
### [1] "TP53" "MYCN"
}
}
