% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-gene-ids.R
\name{ensembl_to_symbol}
\alias{ensembl_to_symbol}
\title{Convert Ensembl gene ids to gene symbols based on reference gene lists}
\usage{
ensembl_to_symbol(
  ensembl_ids,
  reference = c("scpca", "10x2020", "10x2024"),
  sce = NULL,
  unique = FALSE,
  leave_na = FALSE,
  seurat_compatible = FALSE
)
}
\arguments{
\item{ensembl_ids}{A character vector of Ensembl gene ids to translate to
gene symbols.}

\item{reference}{The reference gene list to use for translation. One of
`scpca`, `10x2020`, `10x2024`. The `scpca` reference is the default.}

\item{sce}{A SingleCellExperiment object to use as a reference for gene
symbols. If provided, the `reference` argument will be ignored. The `sce`
object must include columns with the names `gene_ids` (containing Ensembl
ids) and `gene_symbol` (containing the symbols) to use for conversion.}

\item{unique}{Whether to use unique gene symbols, as would be done if data
had been read in with gene symbols by Seurat. Default is FALSE.}

\item{leave_na}{Whether to leave NA values in the output vector. If FALSE,
any missing values will be replaced with the input ensembl_id value.
Default is FALSE.}

\item{seurat_compatible}{Whether to return a vector that is compatible with
Seurat, translating and underscores to dashes.
Default is FALSE.}
}
\value{
A vector of gene symbols corresponding to the input Ensembl ids.
}
\description{
The SingleCellExperiment objects produced as part of ScPCA are indexed by
Ensembl gene ids, as those are more stable than gene symbols. However,
for many applications gene symbols are useful. This function provides
simple conversion of Ensembl gene ids to gene symbols based on either the
ScPCA reference gene list or a 10x reference gene list as used by Cell Ranger.
Alternatively, a SingleCellExperiment object with gene ids and gene symbols
stored in the row data (as those provided by ScPCA) can be used as the reference.
}
\details{
The gene symbols can either be made unique (as would be done if read in by Seurat)
or left as is.
}
\examples{
# convert a set of Ensembl ids to gene symbols
ensembl_ids <- c("ENSG00000141510", "ENSG00000134323")
gene_symbols <- ensembl_to_symbol(ensembl_ids)
gene_symbols
### [1] "TP53" "MYCN"

# convert a set of Ensembl ids to gene symbols using the 10x2020 reference
gene_symbols_10x2020 <- ensembl_to_symbol(ensembl_ids, reference = "10x2020")

\dontrun{
# convert a set of Ensembl ids to gene symbols using an SCE for reference
gene_symbols_sce <- ensembl_to_symbol(ensembl_ids, sce = sce)
}

}
