% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-clusters.R
\name{extract_pc_matrix}
\alias{extract_pc_matrix}
\title{Extract a principal components (PC) matrix from either a SingleCellExperiment
or a Seurat object.}
\usage{
extract_pc_matrix(sc_object, object_type, pc_name)
}
\arguments{
\item{sc_object}{Either a SingleCellExperiment or Seurat object}

\item{object_type}{Whether the object is a SingleCellExperiment ("sce") or a Seurat ("seurat") object}

\item{pc_name}{The name of the PC matrix in the object}
}
\value{
PC matrix with row names
}
\description{
This function extract the PC matrix from either a SingleCellExperiment or Seurat object, as
determined by the object_type argument.
}
