% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum-duplicate-genes.R
\name{sum_duplicate_genes}
\alias{sum_duplicate_genes}
\title{Sum counts for genes with duplicate names in a SingleCellExperiment object.}
\usage{
sum_duplicate_genes(sce, normalize = TRUE, recalculate_reduced_dims = FALSE)
}
\arguments{
\item{sce}{a SingleCellExperiment object with duplicated row names}

\item{normalize}{a logical indicating whether to normalize the expression
values. Default is TRUE.}

\item{recalculate_reduced_dims}{a logical indicating whether to recalculate
PCA and UMAP. If FALSE, the input reduced dimensions are copied over. If
TRUE, the highly variable genes are also recalculated with the new values
stored in metadata. Default is FALSE.}
}
\value{
a SingleCellExperiment object
}
\description{
Genes with the same name are merged by summing their raw expression counts.
When multiple Ensembl gene IDs are associated with the same gene symbol,
identifier conversion can result in duplicate gene names. This function
resolves such duplicates by summing the expression values for each duplicate
gene name, which may be justified if the different Ensembl gene IDs share
substantial sequence identity, which could make separate quantification of
the two genes less reliable.
}
\details{
If requested, the log-normalized expression values are recalculated,
otherwise that matrix is left blank.

PCA and UMAP are also recalculated if requested (which requires recalculating
the normalized expression). If not recalculating, the original reduced
dimensions are retained as they are likely to remain similar in most cases.
}
\examples{
\dontrun{
# sum across duplicated gene names
summed_sce <- sum_duplicate_genes(sce)

# sum across duplicated gene names and recalculate PCA and UMAP
summed_sce <- sum_duplicate_genes(sce, recalculate_reduced_dims = TRUE)
}

}
